+++ src/packages/media/media/index.ts
@@ -9,7 +9,7 @@ export * from './utils/index.js';
 
 export { UMB_MEDIA_TREE_ALIAS, UMB_MEDIA_TREE_PICKER_MODAL } from './tree/index.js';
 export { UMB_MEDIA_COLLECTION_ALIAS } from './collection/index.js';
-export { UMB_MEDIA_MENU_ALIAS } from './menu/manifests.js';
+export { UMB_MEDIA_MENU_ALIAS } from './menu/index.js';
 export { UMB_MEDIA_PICKER_MODAL } from './modals/media-picker/index.js';
 
 export type { UmbMediaTreeItemModel } from './tree/index.js';
