+++ src/packages/media/media/menu/constants.ts
@@ -1,0 +1,1 @@
+export const UMB_MEDIA_MENU_ALIAS = 'Umb.Menu.Media';
