+++ src/packages/media/media/menu/manifests.ts
@@ -1,7 +1,6 @@
 import { UMB_MEDIA_TREE_ALIAS } from '../tree/index.js';
+import { UMB_MEDIA_MENU_ALIAS } from './constants.js';
 import type { ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
-
-export const UMB_MEDIA_MENU_ALIAS = 'Umb.Menu.Media';
 
 export const manifests: Array<ManifestTypes> = [
 	{
