+++ src/packages/media/media/repository/detail/constants.ts
@@ -1,0 +1,2 @@
+export const UMB_MEDIA_DETAIL_REPOSITORY_ALIAS = 'Umb.Repository.Media.Detail';
+export const UMB_MEDIA_DETAIL_STORE_ALIAS = 'Umb.Store.Media.Detail';
