+++ src/packages/media/media/repository/detail/index.ts
@@ -1,2 +1,2 @@
 export { UmbMediaDetailRepository } from './media-detail.repository.js';
-export { UMB_MEDIA_DETAIL_REPOSITORY_ALIAS } from './manifests.js';
+export { UMB_MEDIA_DETAIL_REPOSITORY_ALIAS } from './constants.js';
