+++ src/packages/media/media/repository/detail/manifests.ts
@@ -1,6 +1,5 @@
+import { UMB_MEDIA_DETAIL_REPOSITORY_ALIAS, UMB_MEDIA_DETAIL_STORE_ALIAS } from './constants.js';
 import type { ManifestRepository, ManifestStore, ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
-
-export const UMB_MEDIA_DETAIL_REPOSITORY_ALIAS = 'Umb.Repository.Media.Detail';
 
 const repository: ManifestRepository = {
 	type: 'repository',
@@ -8,8 +7,6 @@ const repository: ManifestRepository = {
 	name: 'Media Detail Repository',
 	api: () => import('./media-detail.repository.js'),
 };
-
-export const UMB_MEDIA_DETAIL_STORE_ALIAS = 'Umb.Store.Media.Detail';
 
 const store: ManifestStore = {
 	type: 'store',
